Updated ch3-poly-regression.py.

+++ maning_shukla/ch3-poly-regression.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 learning_rate = 0.01
 training_epochs = 40
 
-trX = np.linspace(-1, 1, 51)
+trX = np.linspace(-1, 1, 51, dtype = "float")
 
 
 num_coeffs = 6
